# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "version": "1.0.0",
      "title": "Contacts App",
      "license": {
        "name": "Apache 2.0",
        "url": "<http://www.apache.org/licenses/LICENSE-2.0.html>"
      },
      "description": "This is a documentation of contacts app\n"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      },
      {
        "url": "https://nodejs-hw-mongodb-gue9.onrender.com"
      }
    ],
    "tags": [
      {
        "name": "Contacts",
        "description": "Operations about contacts."
      },
      {
        "name": "Auth",
        "description": "Auth operations."
      }
    ],
    "paths": {
      "/contacts": {
        "get": {
          "tags": [
            "Contacts"
          ],
          "summary": "Get list of contacts",
          "operationId": "getContacts",
          "description": "Get list of contacts",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "query",
              "name": "page",
              "schema": {
                "type": "integer"
              },
              "example": 1
            },
            {
              "in": "query",
              "name": "perPage",
              "schema": {
                "type": "integer",
                "example": 10
              }
            },
            {
              "in": "query",
              "name": "sortBy",
              "schema": {
                "type": "string",
                "example": "name"
              },
              "description": "All contacts's fields can be mentioned. Default value - _id"
            },
            {
              "in": "query",
              "name": "sortOrder",
              "schema": {
                "type": "string",
                "enum": [
                  "asc",
                  "desc"
                ],
                "example": "asc"
              }
            },
            {
              "in": "query",
              "name": "isFavourite",
              "schema": {
                "type": "boolean",
                "example": true
              }
            },
            {
              "in": "query",
              "name": "contactType",
              "schema": {
                "type": "string",
                "enum": [
                  "work",
                  "home",
                  "personal"
                ],
                "example": "work"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successfully found contacts!",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully found contacts!"
                      },
                      "data": {
                        "type": "object",
                        "required": [
                          "data",
                          "page",
                          "perPage",
                          "totalItems",
                          "totalPages",
                          "hasNextPage",
                          "hasPreviousPage"
                        ],
                        "properties": {
                          "data": {
                            "type": "array",
                            "properties": {
                              "data": {
                                "$ref": "#/components/schemas/contact"
                              }
                            }
                          },
                          "page": {
                            "type": "number",
                            "example": 1
                          },
                          "perPage": {
                            "type": "number",
                            "example": 5
                          },
                          "totalItems": {
                            "type": "number",
                            "example": 4
                          },
                          "totalPages": {
                            "type": "number",
                            "example": 4
                          },
                          "hasNextPage": {
                            "type": "boolean",
                            "example": false
                          },
                          "hasPreviousPage": {
                            "type": "boolean",
                            "example": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        },
        "post": {
          "tags": [
            "Contacts"
          ],
          "summary": "Create contact",
          "operationId": "createContact",
          "description": "Create a contact with payload",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name",
                    "phoneNumber",
                    "contactType",
                    "userId"
                  ],
                  "properties": {
                    "name": {
                      "description": "Contact's name",
                      "type": "string",
                      "example": "Andriana Petrenko"
                    },
                    "phoneNumber": {
                      "description": "Contact's number",
                      "type": "string",
                      "example": 380963251212
                    },
                    "email": {
                      "description": "Contact's email",
                      "type": "string",
                      "example": "tajo@gmail.com"
                    },
                    "contactType": {
                      "description": "Contact's type",
                      "type": "string",
                      "enum": [
                        "work",
                        "home",
                        "personal"
                      ]
                    },
                    "isFavourite": {
                      "description": "Whether is contact favourite",
                      "type": "boolean",
                      "example": false
                    },
                    "userId": {
                      "description": "Contact's user id",
                      "type": "string",
                      "example": "65e4decdd286b30065d54af9"
                    },
                    "photo": {
                      "description": "Contact's photo",
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Creates a student",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 201
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully created a contact!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/contact"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      },
      "/contacts/{contactId}": {
        "get": {
          "tags": [
            "Contacts"
          ],
          "summary": "Get contact by id",
          "operationId": "getContact",
          "description": "Get contact by id",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "contactId",
              "required": true,
              "schema": {
                "type": "string",
                "example": "6675a7235e9619e247d16c13"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Get contact by id",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully found contact with id 6675a7235e9619e247d16c13!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/contact"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        },
        "delete": {
          "tags": [
            "Contacts"
          ],
          "summary": "Delete contact by id",
          "operationId": "deleteContact",
          "description": "Delete contact by id",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "contactId",
              "required": true,
              "schema": {
                "type": "string",
                "example": "6675a7235e9619e247d16c13"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Delete contact by id"
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        },
        "patch": {
          "tags": [
            "Contacts"
          ],
          "summary": "Patch contact",
          "operationId": "patchContact",
          "description": "Update contact",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "contactId",
              "required": true,
              "schema": {
                "type": "string",
                "example": "6675a7235e9619e247d16c13"
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "description": "Contact's name",
                      "type": "string",
                      "example": "Andriana Petrenko"
                    },
                    "phoneNumber": {
                      "description": "Contact's number",
                      "type": "string",
                      "example": 380963251212
                    },
                    "email": {
                      "description": "Contact's email",
                      "type": "string",
                      "example": "tajo@gmail.com"
                    },
                    "contactType": {
                      "description": "Contact's type",
                      "type": "string",
                      "enum": [
                        "work",
                        "home",
                        "personal"
                      ]
                    },
                    "isFavourite": {
                      "description": "Whether is contact favourite",
                      "type": "boolean",
                      "example": false
                    },
                    "userId": {
                      "description": "Contact's user id",
                      "type": "string",
                      "example": "65e4decdd286b30065d54af9"
                    },
                    "photo": {
                      "description": "Contact's photo",
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Update contact by id",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully patched a contact!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/contact"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        }
      },
      "/auth/register": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Register new user",
          "operationId": "createUser",
          "description": "Create a user with payload",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name",
                    "email",
                    "password"
                  ],
                  "properties": {
                    "name": {
                      "description": "User's name",
                      "type": "string",
                      "example": "Andriana Petrenko"
                    },
                    "email": {
                      "description": "User email",
                      "type": "string",
                      "example": "tajo@gmail.com"
                    },
                    "password": {
                      "description": "User's password",
                      "type": "string",
                      "example": "65e4decdd286"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Creates a new user",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 201
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully registered a user!"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "description": "User's id",
                            "type": "string",
                            "example": "65e4decdd286b30065d54af9"
                          },
                          "name": {
                            "description": "User's name",
                            "type": "string",
                            "example": "Oleg"
                          },
                          "email": {
                            "description": "User's email",
                            "type": "string",
                            "example": "tajo@gmail.com"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "409": {
              "$ref": "#/components/responses/409"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Login user",
          "operationId": "loginUser",
          "description": "Login a user",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "email",
                    "password"
                  ],
                  "properties": {
                    "email": {
                      "description": "User email",
                      "type": "string",
                      "example": "tajo@gmail.com"
                    },
                    "password": {
                      "description": "User's password",
                      "type": "string",
                      "example": "65e4decdd286"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login a user",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully logged in an user"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "accessToken": {
                            "type": "string",
                            "example": "Pw/Tu2KKtZI4jsPjP1Tklym3yFsUERU1EV3HM37y200"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      },
      "/auth/logout": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Logout user",
          "operationId": "logoutUser",
          "description": "Logout a user",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "cookie",
              "name": "sessionId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "The session ID from the cookie",
              "example": "j%3A%22667fbc6aca1f959e6f4efacf%22"
            },
            {
              "in": "cookie",
              "name": "refreshToken",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "The refresh token from the cookie",
              "example": "hfP76WlQAwMI7Dzh%2BbRBC6m%2FzIvY96Y7awZWbBHu"
            }
          ],
          "responses": {
            "204": {
              "description": "Successfully logged out the user"
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      },
      "/auth/send-reset-email": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Send reset email",
          "operationId": "sendEmail",
          "description": "Send reset password email",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "email"
                  ],
                  "properties": {
                    "email": {
                      "description": "User email",
                      "type": "string",
                      "example": "tajo@gmail.com"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Send reset password email",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Reset password email was successfully sent!"
                      },
                      "data": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        }
      },
      "/auth//reset-pwd": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Reset password",
          "operationId": "resetPassword",
          "description": "Reset password",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "token",
                    "password"
                  ],
                  "properties": {
                    "token": {
                      "description": "User token",
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NjcyZjA5Y2NhMWQ5OTZlNTNmZDg5OTIiLCJlbWFpbCI6InpsYXRhMDcwN3BldHJlbmtvQGdtYWlsLmNvbSIsImlhdCI6MTcxODk5MzUzNiwiZXhwIjoxNzE4OTkzODM2fQ.WKG8p4l-DT6QYVcBTHLfHMIjhuwRA1XOLGF4PEooffQ"
                    },
                    "password": {
                      "description": "User's password",
                      "type": "string",
                      "example": "65e4decdd286"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Reset password",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Password was successfully reset!"
                      },
                      "data": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        }
      },
      "/auth//refresh": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Refresh user session",
          "operationId": "refreshSession",
          "description": "Refreshes the user session",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "cookie",
              "name": "sessionId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "The session ID from the cookie",
              "example": "j%3A%22667fbc6aca1f959e6f4efacf%22"
            },
            {
              "in": "cookie",
              "name": "refreshToken",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "The refresh token from the cookie",
              "example": "hfP76WlQAwMI7Dzh%2BbRBC6m%2FzIvY96Y7awZWbBHu"
            }
          ],
          "responses": {
            "200": {
              "description": "Successfully refreshed a session",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully refreshed a session!"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "accessToken": {
                            "type": "string",
                            "description": "The new access token",
                            "example": "tkntaXI2mXcEBe8tMjBYsvrJZeTUse7J+SJSNS2z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      },
      "schemas": {
        "contact": {
          "type": "object",
          "required": [
            "name",
            "phoneNumber",
            "contactType",
            "userId"
          ],
          "properties": {
            "_id": {
              "description": "Contact's id",
              "type": "string",
              "example": "65e4decdd286b30065d54af9"
            },
            "name": {
              "description": "Contact's name",
              "type": "string",
              "example": "Andriana Petrenko"
            },
            "phoneNumber": {
              "description": "Contact's number",
              "type": "string",
              "example": 380963251212
            },
            "email": {
              "description": "Contact's email",
              "type": "string",
              "example": "tajo@gmail.com"
            },
            "contactType": {
              "description": "Contact's type",
              "type": "string",
              "enum": [
                "work",
                "home",
                "personal"
              ]
            },
            "isFavourite": {
              "description": "Whether is contact favourite",
              "type": "boolean",
              "example": false
            },
            "userId": {
              "description": "Contact's user id",
              "type": "string",
              "example": "65e4decdd286b30065d54af9"
            },
            "photo": {
              "description": "Link to contact's photo",
              "type": "string",
              "example": "<https://res.cloudinary.com/uqwgdu/image/upload/image.png>"
            }
          }
        }
      },
      "responses": {
        "401": {
          "description": "Unauthorized error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 401
                  },
                  "message": {
                    "type": "string",
                    "example": "UnauthorizedError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Token is expired or invalid"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "Unauthorized error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 404
                  },
                  "message": {
                    "type": "string",
                    "example": "UnauthorizedError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "409": {
          "description": "Conflict error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 409
                  },
                  "message": {
                    "type": "string",
                    "example": "ConflictError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Email is in use"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }